172721 - Code refactored AB#172721 AB#176647
