+++ src/UKHO.S100PermitService.API/appsettings.json
@@ -44,7 +44,7 @@
     "DeductTokenExpiryMinutes": 1
   },
   "HoldingsServiceApiConfiguration": {
-    "HoldingsClientId": "",
+    "ClientId": "",
     "BaseUrl": ""
   }
 }
