+++ src/UKHO.S100PermitService.Common/Configuration/HoldingsServiceApiConfiguration.cs
@@ -2,7 +2,7 @@
 {
     public class HoldingsServiceApiConfiguration : IHoldingsServiceApiConfiguration
     {
-        public string HoldingsClientId { get; set; }
+        public string ClientId { get; set; }
         public string BaseUrl { get; set; }
     }
 }
