+++ src/UKHO.S100PermitService.Common/Configuration/IHoldingsServiceApiConfiguration.cs
@@ -2,7 +2,7 @@
 {
     public interface IHoldingsServiceApiConfiguration
     {
-        public string HoldingsClientId { get; set; }
+        public string ClientId { get; set; }
         public string BaseUrl { get; set; }
     }
 }
